requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-filter==23.2
django-multiupload==0.6.1
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
pillow==9.5.0
djangorestframework-simplejwt==5.2.2
django_rest_passwordreset==1.3.0
psycopg2-binary==2.9.6
